feat: redis 적용 테스트 api 추가
