+++ momentia-api/src/main/java/org/ject/momentia/api/artwork/controller/ArtworkController.java
@@ -13,8 +13,6 @@ import org.ject.momentia.api.mvc.annotation.EnumValue;
 import org.ject.momentia.api.mvc.annotation.MomentiaUser;
 import org.ject.momentia.api.pagination.model.PaginationResponse;
 import org.ject.momentia.api.pagination.model.PaginationWithIsMineResponse;
-import org.ject.momentia.api.test.testCacheModel;
-import org.ject.momentia.api.test.testCacheRepository;
 import org.ject.momentia.common.domain.artwork.type.Category;
 import org.ject.momentia.common.domain.user.User;
 import org.springframework.http.HttpStatus;
@@ -41,7 +39,6 @@ import lombok.RequiredArgsConstructor;
 public class ArtworkController {
 
 	private final ArtworkPostService artworkPostService;
-	private final testCacheRepository testRepository;
 
 	/**
 	 * [POST] 작품 업로드
@@ -173,50 +170,4 @@ public class ArtworkController {
 		return artworkPostService.getLikePosts(sort, page, size, user);
 	}
 
-	/**
-	 * 테스트 용 API - 추후 삭제
-	 */
-	@GetMapping("/redis/{num}")
-	@ResponseStatus(HttpStatus.OK)
-	public String getTestCache(@PathVariable Long num) {
-		if (testRepository.findById(num).isPresent()) {
-			return "기존 - " + testRepository.findById(num).get().getName();
-		} else
-			return "새로 생성 - " + testRepository.save(new testCacheModel(num, "name" + num)).getName();
-	}
-
-	/**
-	 * 테스트 용 API - 추후 삭제
-	 */
-	@GetMapping("/redis/{num}/delete")
-	@ResponseStatus(HttpStatus.OK)
-	public String deleteTestCache(@PathVariable Long num) {
-		if (testRepository.findById(num).isEmpty()) {
-			return "없음";
-		} else {
-			testRepository.deleteById(num);
-			return "삭제 완료";
-		}
-	}
-
-	/**
-	 * 테스트 용 API - 추후 삭제
-	 */
-	@GetMapping("/redis/deleteAll")
-	@ResponseStatus(HttpStatus.OK)
-	public String deleteAllTestCache() {
-		testRepository.deleteAll();
-		return "모두 삭제";
-	}
-
-	@GetMapping("/redis/findAll")
-	@ResponseStatus(HttpStatus.OK)
-	public String findAllCache() {
-		Iterable<testCacheModel> list = testRepository.findAll();
-		StringBuilder all = new StringBuilder();
-		for (testCacheModel model : list) {
-			all.append(model.getName());
-		}
-		return all.toString();
-	}
 }
